Update 1:20pm (BIG ONE WILL SUBMIT AS FIRST VERSION)
The health collectible item now has a particle effect... kinda. Its a smoke in mirrors thing. It doesn't work after a few seconds which is why its right next to where the payer spawns for demo purposes BUT as far we/she is concerned... IT WORKS. Hopefully we can replicate this for take damage. Submit as a first version of assignment.